final_draft/requirements.txt
flask==3.0.2
opencv-python==4.8.0.74
mediapipe==0.10.5
numpy==1.24.3
deep-translator==1.11.4
textblob==0.17.1
pyttsx3==2.90
waitress==3.0.0
gunicorn==21.2.0
customtkinter==5.2.0
Pillow==10.0.0
